build script checks python version

+++ build.py
@@ -1,6 +1,12 @@
 import os
 import shutil
 import subprocess
+import sys
+
+if sys.version_info < (3,5):
+    print("Need python3 to build. Exiting")
+    sys.exit(1)
+
 
 try:
     shutil.rmtree("dist/")
